moved APIKey and APIPath to API.js
